requirements.txt
feedparser
newspaper3k
pywhatkit
apscheduler
twilio
PyJSONViewer
psycopg2
selenium==3.141.0
kivy
numpy
python-Levenshtein
fuzzywuzzy